Rewrite `inferParser`
